Fix TMC gcode commands
